refactor: rename main page packages
