+++ src/components/DeviceCard/DeviceCard.module.css
