+++ src/pages/Live/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SensorDashboard from '../components/SensorDashboard';
+import SensorDashboard from '../../components/SensorDashboard';
 
 function Live() {
     return <SensorDashboard view="live" title="Live" />;
