+++ src/pages/Reports/index.jsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import Header from '../components/Header';
-import { useLiveDevices } from '../components/dashboard/useLiveDevices';
-import { useHistory } from '../components/dashboard/useHistory';
-import styles from '../components/SensorDashboard.module.css';
-import ReportControls from '../components/dashboard/ReportControls';
-import ReportCharts from '../components/dashboard/ReportCharts';
-import { SENSOR_TOPIC, topics } from '../components/dashboard/dashboard.constants';
-import { toLocalInputValue, formatTime } from '../components/dashboard/dashboard.utils';
-import { useFilters, ALL } from '../context/FiltersContext';
+import Header from '../../components/Header';
+import { useLiveDevices } from '../../components/dashboard/useLiveDevices';
+import { useHistory } from '../../components/dashboard/useHistory';
+import styles from '../../components/SensorDashboard/SensorDashboard.module.css';
+import ReportControls from '../../components/dashboard/ReportControls';
+import ReportCharts from '../../components/dashboard/ReportCharts';
+import { SENSOR_TOPIC, topics } from '../../components/dashboard/dashboard.constants';
+import { toLocalInputValue, formatTime } from '../../components/dashboard/dashboard.utils';
+import { useFilters, ALL } from '../../context/FiltersContext';
 
-function ReportsPage() {
+function Reports() {
     const [activeSystem, setActiveSystem] = useState('S01');
     const { deviceData, availableCompositeIds } = useLiveDevices(topics, activeSystem);
     const [selectedDevice, setSelectedDevice] = useState('');
@@ -187,5 +187,5 @@ function ReportsPage() {
     );
 }
 
-export default ReportsPage;
+export default Reports;
